Added menus, box pushing, and winning
The loadNewLevel also now calls a loadFromBackup void to search through
backupLevel for boxes and the player spawn. This loadFromBackup is also
called when the level is reset.

I have noticed an issue that the loadNewLevel always seems to load only
the first level; I'm assuming this is because it closes the Scanner.